Change AbstractValueObject to type
Change the abstract AbstractValueObject object to the
AbstractConditionalType type, because types should be classes which have
only one value and value objects should be classes which have one or
more values but behave every time in the same way. So the type wrapper
class for conditional types is more a type then already a value object.